+++ storybook/src/components/PageFooter/PageFooter.stories.tsx
@@ -68,7 +68,7 @@ export const Default: Story = {
               Werken bij Amsterdam
             </StandaloneLink>
           </Grid.Cell>
-          <Grid.Cell span={4} start={{ narrow: 1, medium: 5, wide: 5 }}>
+          <Grid.Cell span={4}>
             <Heading className="ams-mb-s" color="inverse" level={2} size="level-3">
               Volg ons
             </Heading>
@@ -80,7 +80,7 @@ export const Default: Story = {
               ))}
             </LinkList>
           </Grid.Cell>
-          <Grid.Cell span={4} start={{ narrow: 1, medium: 1, wide: 9 }}>
+          <Grid.Cell span={4}>
             <Heading className="ams-mb-s" color="inverse" level={2} size="level-3">
               Doen in de stad
             </Heading>
@@ -98,10 +98,7 @@ export const Default: Story = {
           </Grid.Cell>
         </Grid>
       </PageFooter.Spotlight>,
-      <Heading className="ams-visually-hidden" key={2} level={2}>
-        Over deze website
-      </Heading>,
-      <PageFooter.Menu key={3}>
+      <PageFooter.Menu key={2}>
         <PageFooter.MenuLink href="#">Over deze site</PageFooter.MenuLink>
         <PageFooter.MenuLink href="#">Privacy</PageFooter.MenuLink>
         <PageFooter.MenuLink href="#">Cookies op deze site</PageFooter.MenuLink>
@@ -116,7 +113,7 @@ export const OnderzoekEnStatistiek: Story = {
     children: [
       <PageFooter.Spotlight key={1}>
         <Grid paddingVertical="x-large">
-          <Grid.Cell span={3}>
+          <Grid.Cell span={4}>
             <Heading className="ams-mb-s" color="inverse" level={2} size="level-3">
               Contact
             </Heading>
@@ -132,7 +129,7 @@ export const OnderzoekEnStatistiek: Story = {
               </LinkList.Link>
             </LinkList>
           </Grid.Cell>
-          <Grid.Cell span={3} start={{ narrow: 1, medium: 5, wide: 5 }}>
+          <Grid.Cell span={4}>
             <Heading className="ams-mb-s" color="inverse" level={2} size="level-3">
               Panels en enquêtes
             </Heading>
@@ -151,7 +148,7 @@ export const OnderzoekEnStatistiek: Story = {
               </LinkList.Link>
             </LinkList>
           </Grid.Cell>
-          <Grid.Cell span={3} start={{ narrow: 1, medium: 1, wide: 9 }}>
+          <Grid.Cell span={4}>
             <Heading className="ams-mb-s" color="inverse" level={2} size="level-3">
               Onderzoek en Statistiek
             </Heading>
@@ -175,13 +172,23 @@ export const OnderzoekEnStatistiek: Story = {
           </Grid.Cell>
         </Grid>
       </PageFooter.Spotlight>,
-      <Heading className="ams-visually-hidden" key={2} level={2}>
-        Over deze website
-      </Heading>,
-      <PageFooter.Menu key={3}>
+      <PageFooter.Menu key={2}>
         <PageFooter.MenuLink href="#">Privacy</PageFooter.MenuLink>
         <PageFooter.MenuLink href="#">Toegankelijkheid</PageFooter.MenuLink>
       </PageFooter.Menu>,
     ],
+  },
+}
+
+export const CustomMenuHeading: Story = {
+  args: {
+    children: (
+      <PageFooter.Menu heading="Support links">
+        <PageFooter.MenuLink href="#">About this website</PageFooter.MenuLink>
+        <PageFooter.MenuLink href="#">Privacy</PageFooter.MenuLink>
+        <PageFooter.MenuLink href="#">Cookies</PageFooter.MenuLink>
+        <PageFooter.MenuLink href="#">Web archive</PageFooter.MenuLink>
+      </PageFooter.Menu>
+    ),
   },
 }