+++ storybook/src/pages/public/ArticlePage/ArticlePage.stories.tsx
@@ -1,0 +1,28 @@
+/**
+ * @license EUPL-1.2+
+ * Copyright Gemeente Amsterdam
+ */
+
+import type { Meta, StoryObj } from '@storybook/react-vite'
+
+import type { NewsArticle } from './news-articles'
+
+import { commonMeta } from '../common/config'
+import { ArticlePage } from './ArticlePage'
+import { newsArticles } from './news-articles'
+
+const meta = {
+  ...commonMeta,
+  title: 'Pages/Public/Article Page',
+  component: ArticlePage,
+} satisfies Meta<typeof ArticlePage>
+
+export default meta
+
+const getArgsFromNewsArticle = (id: number): NewsArticle | undefined => {
+  return newsArticles.find((article) => article.id === id)
+}
+
+export const Default: StoryObj = {
+  args: getArgsFromNewsArticle(1),
+}