+++ storybook/src/pages/public/ProductPage/ProductPage.stories.tsx
@@ -1,0 +1,19 @@
+/**
+ * @license EUPL-1.2+
+ * Copyright Gemeente Amsterdam
+ */
+
+import type { Meta, StoryObj } from '@storybook/react-vite'
+
+import { commonMeta } from '../common/config'
+import { ProductPage } from './ProductPage'
+
+const meta = {
+  ...commonMeta,
+  title: 'Pages/Public/Product Page',
+  component: ProductPage,
+} satisfies Meta<typeof ProductPage>
+
+export default meta
+
+export const Default: StoryObj = {}