+++ storybook/src/pages/public/common/AppHeader.tsx
@@ -1,0 +1,51 @@
+/**
+ * @license EUPL-1.2+
+ * Copyright Gemeente Amsterdam
+ */
+
+import { Grid, Heading, LinkList, PageHeader } from '@amsterdam/design-system-react'
+
+import { megaMenuLinks, pageHeaderMenuLinks } from './menu'
+
+export const AppHeader = () => (
+  <PageHeader
+    menuItems={pageHeaderMenuLinks.map(({ fixed, href, icon, label, lang }) => (
+      <PageHeader.MenuLink
+        fixed={fixed}
+        href={href ?? '#'}
+        icon={icon}
+        key={label}
+        lang={lang}
+        rel={href ? 'external' : undefined}
+      >
+        {label}
+      </PageHeader.MenuLink>
+    ))}
+  >
+    <Grid>
+      <PageHeader.GridCellNarrowWindowOnly span="all">
+        <LinkList>
+          {pageHeaderMenuLinks
+            .filter((link) => !link.fixed)
+            .map(({ href, icon, label, lang }) => (
+              <LinkList.Link href={href ?? '#'} icon={icon} key={label} lang={lang} rel={href ? 'external' : undefined}>
+                {label}
+              </LinkList.Link>
+            ))}
+        </LinkList>
+      </PageHeader.GridCellNarrowWindowOnly>
+      <Grid.Cell span="all">
+        <Heading className="ams-mb-s" level={2}>
+          Alle onderwerpen
+        </Heading>
+        <LinkList>
+          {megaMenuLinks.map((label) => (
+            <LinkList.Link href="#" key={label}>
+              {label}
+            </LinkList.Link>
+          ))}
+        </LinkList>
+      </Grid.Cell>
+    </Grid>
+  </PageHeader>
+)